+++ Biwen.Settings/SettingStores/SettingRecord.cs
@@ -3,17 +3,8 @@
 // See the LICENSE file in the project root for more information.
 // Biwen.Settings Author: 万雅虎, Github: https://github.com/vipwan
 // Biwen.Settings ,NET8+ 应用配置项管理模块
-// Modify Date: 2024-09-18 17:29:58 EFCoreStoreOptions.cs
+// Modify Date: 2024-09-18 17:30:45 SettingRecord.cs
 
-using static Biwen.Settings.SettingOptions;
+namespace Biwen.Settings.SettingStores;
 
-namespace Biwen.Settings.SettingManagers.EFCore;
-
-public class EFCoreStoreOptions
-{
-    /// <summary>
-    /// 加密的内容配置项
-    /// </summary>
-    public EncryptionOptions EncryptionOptions { get; set; } = new();
-
-}
+internal record SettingRecord(Type SettingType, object Setting);